Remove the InMemoryTenantRepository implementation
We ended up not using it for now.
