--- Services/CO.CDP.Tenant.Persistence.Tests/InMemoryTenantRepositoryTest.cs
@@ -1,9 +1,0 @@
-namespace CO.CDP.Tenant.Persistence.Tests;
-
-public class InMemoryTenantRepositoryTest : TenantRepositoryContractTest
-{
-    protected override ITenantRepository TenantRepository()
-    {
-        return new InMemoryTenantRepository();
-    }
-}
